Make remove event parameters match change event parameters.
A change event should show the new value of the viewed attribute. For
example, 'change:root.depth1' should send as its second parameter the
new depth1 value.

This should still be the case when removing a value from an array; that
is, the second parameter of the change event should be consistent,
regardless of the operation that caused it to change.

Prior to this change, the change event sent when removing an array item
was the removed item.